+++ src/mainNode.ts
@@ -1,28 +1,13 @@
 import { ExtensionContext } from "vscode";
-import { LanguageClient, ServerOptions } from "vscode-languageclient/node";
-import { Coqpilot, ClientFactoryType } from "./extension";
+import { Coqpilot } from "./extension";
 
 let extension: Coqpilot | undefined;
 
-const cf: ClientFactoryType = (_context, clientOptions, wsConfig) => {
-    const serverOptions: ServerOptions = {
-        command: "coq-lsp",
-        args: wsConfig.args,
-    };
-    return new LanguageClient(
-        "coq-lsp",
-        "Coq LSP Server",
-        serverOptions,
-        clientOptions
-    );
-};
-
 export async function activate(context: ExtensionContext): Promise<void> {
-    extension = await Coqpilot.init(context, cf);
+    extension = await Coqpilot.init(context);
     context.subscriptions.push(extension);
 }
   
 export function deactivate() {
-    extension.deactivateCoqLSP();
     extension.dispose();
 }